Fixing KStarsDateTime::epoch().
We had:     epoch = year + double( dayOfYear / daysInYear )
changed to: epoch = year + double( dayOfYear )/double( daysInYear )

CCMAIL: kstars-devel@kde.org

svn path=/trunk/kdeedu/kstars/; revision=332437

+++ kstars/kstarsdatetime.h
@@ -81,7 +81,7 @@ public:
 	QTime GSTtoUT( dms GST ) const;
 	
 	double epoch() const { return ( double( date().year() ) 
-			+ double( date().dayOfYear()/date().daysInYear() ) ); }
+			+ double( date().dayOfYear() )/double( date().daysInYear() ) ); }
 	void setFromEpoch( double e );
 private:
 	dms GSTat0hUT() const;
